requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
scikit-image>=0.21
scikit-learn>=1.3
